Don't manage zuul service state
There are cases where we actually want to restart zuul manually, but we
cannot do that today because our post jobs in zuul will try to enable
the service again, if stopped.  Now we no longer manage that via
windmill, and are free to stop / start zuul services without breaking
the post job.

Depends-On: https://review.openstack.org/631942
Signed-off-by: Paul Belanger <pabelanger@redhat.com>