Manage zuul service states again
So, we have a little issue here. If we bootstrap a replacement server
for the first time, we then need to manually start the server. The fix
here is to likey reboot the server the first time we bootstrap it,
however we don't have automation around that yet.  For now, simply
manage and if we need to stop / start services we can add the server
into the disabled group in the inventory.

Signed-off-by: Paul Belanger <pabelanger@redhat.com>