Add CORS header for log proxy
The Zuul build page uses cross-origin async requests in order to
fetch files from the log server.  Adding this header allows that
to happen.  It means that javascript hosted at any domain is
permitted to request resources from the log server.  That is fine
for public, static resources like this.