disable nested ESXi on Limestone
We still face 50% failure rate on Limestone, whereas the same runs on
Vexxhost are ok.

So, I remove `esxi-6.7.0-with-nested` and create
`esxi-6.7.0-with-nested_unstable`. This will allow the creation of
unstable job that we will use to track the progress we make with
Limestone. This, without impacting our users.