vexxhost: use a pool for the ESXi
The ESXi configuration is now the same between the two Vexxhost zone.
`v2-standard-1-iops` is not available in `sjc1`, and if nodepool retries
to spawn a node there, we get a `node_failure`.

So, it's impossible use the same pool configuration properly. We now
have a pool dedicated to the ESXi nodes.