esxi: ensure fqdn < 63 characters
ESXi FQDN cannot go above 63 characters. By default, the FQDN on
Vexxhost go above this limit. As a result, the FQDN and hostnames
collected by Ansible or `excfg` are actually truncated. This breaks
the name resolution on the hosts.

We don't have the same problem with Limestone, but I prefect to apply
the configuration anyway, this to be consistent.

Also, bump the image to 20200630 to include a version of
esxi-cloud-init.py that can set the hostname properly.