nodepool/f34: use fedora element
The fedora-minimal depends on the yumdownloader command from yum-utils.
The version from Ubuntu 18.04 seems to be too old and not able to
properly resolve the dependency. We always end up with the following
error:
Error: openssl1.1 conflicts with 1:openssl-libs-1.1.1k-1.fc34.x86_64

By using the fedora element instead, we use the official Fedora cloud
image as base image.

We import a copy for the fedora element for now because we need this PR:
https://review.opendev.org/c/openstack/diskimage-builder/+/799339